moved checklist modal into components directory
