+++ src/components/modals/Case/components/Checklist.tsx
@@ -1,9 +1,35 @@
-import React from 'react';
+import { Button, Checkbox, Input } from 'antd';
+import Modal from 'antd/lib/modal/Modal';
+import React, { useState } from 'react';
 
-export default function Checklist() {
+const Checklist = ({ handleCheckboxChange, request }) => {
+  // state for checklist modal
   return (
-    <div>
-      <h1>Checklist</h1>
+    <div
+      style={{
+        display: 'flex',
+        flexDirection: 'column',
+        width: '100%',
+        gap: '0.3rem',
+      }}
+    >
+      <Checkbox
+        checked={request.verifiedDocuments}
+        name="verifiedDocuments"
+        onChange={handleCheckboxChange}
+      >
+        All documents received and verified
+      </Checkbox>
+      <br />
+      <Checkbox
+        checked={request.pmApproval}
+        name="pmApproval"
+        onChange={handleCheckboxChange}
+      >
+        Approved by program manager
+      </Checkbox>
     </div>
   );
-}
+};
+
+export default Checklist;
