+++ src/components/modals/Case/components/index.ts
@@ -1,0 +1,11 @@
+import Basic from './Basic';
+
+import Documents from './Documents';
+
+import Checklist from './Checklist';
+
+import Footer from './Footer';
+
+import TopActions from './TopActions';
+
+export { Basic, Documents, Checklist, Footer, TopActions };
